+++ packages/cds-mobile2/src/controls/__stories__/Checkbox.stories.tsx
@@ -1,4 +1,5 @@
 import React from 'react';
+import { StyleSheet } from 'react-native';
 import { useToggler } from '@cbhq/cds-common2/hooks/useToggler';
 
 import { Example, ExampleScreen } from '../../examples/ExampleScreen';
@@ -6,6 +7,21 @@ import { useCheckboxGroupState } from '../../hooks/useCheckboxGroupState';
 import { TextHeadline } from '../../typography/TextHeadline';
 import { Checkbox } from '../Checkbox';
 import { CheckboxGroup } from '../CheckboxGroup';
+
+const styles = StyleSheet.create({
+  checkboxGroup: {
+    flexDirection: 'row',
+    flexWrap: 'wrap',
+    gap: 8,
+    marginTop: 8,
+  },
+});
+
+const options = {
+  'fish-taco': 'Fish tacos',
+  puttanesca: 'Spaghetti alla puttanesca',
+  'hamachi-salad': 'Hamachi salad',
+};
 
 const CheckboxScreen = () => {
   return (
@@ -40,12 +56,6 @@ const CheckboxScreen = () => {
 
       <Example inline title="Checkbox Group">
         {() => {
-          const options = {
-            'fish-taco': 'Fish tacos',
-            puttanesca: 'Spaghetti alla puttanesca',
-            'hamachi-salad': 'Hamachi salad',
-          };
-
           const optionValues = Object.keys(options) as unknown as (keyof typeof options)[];
           const [selectedValues, { toggle }] =
             // eslint-disable-next-line react-hooks/rules-of-hooks
@@ -58,6 +68,33 @@ const CheckboxScreen = () => {
                 accessibilityLabel="Order Dinner"
                 onChange={toggle}
                 selectedValues={selectedValues}
+              >
+                {Object.entries(options).map(([value, label]) => (
+                  <Checkbox key={value} value={value}>
+                    {label}
+                  </Checkbox>
+                ))}
+              </CheckboxGroup>
+            </>
+          );
+        }}
+      </Example>
+
+      <Example inline title="Checkbox Group With Styles">
+        {() => {
+          const optionValues = Object.keys(options) as unknown as (keyof typeof options)[];
+          const [selectedValues, { toggle }] =
+            // eslint-disable-next-line react-hooks/rules-of-hooks
+            useCheckboxGroupState<keyof typeof options>(optionValues);
+
+          return (
+            <>
+              <TextHeadline>Order Dinner</TextHeadline>
+              <CheckboxGroup<keyof typeof options>
+                accessibilityLabel="Order Dinner"
+                onChange={toggle}
+                selectedValues={selectedValues}
+                style={styles.checkboxGroup}
               >
                 {Object.entries(options).map(([value, label]) => (
                   <Checkbox key={value} value={value}>