+++ packages/cds-web2/src/controls/CheckboxGroup.tsx
@@ -6,7 +6,7 @@ import React, {
   isValidElement,
   memo,
 } from 'react';
-import { css } from '@linaria/core';
+import { css, cx } from '@linaria/core';
 import type { SharedProps } from '@cbhq/cds-common2/types';
 import type { CheckboxGroupBaseProps } from '@cbhq/cds-common2/types/CheckboxGroupBaseProps';
 import { isDevelopment } from '@cbhq/cds-utils';
@@ -23,8 +23,10 @@ const checkboxStyles = css`
 `;
 
 export type CheckboxGroupProps<T extends string> = {
+  className?: string;
   /** Handle change event when pressing on a checkbox option. */
   onChange?: React.ChangeEventHandler<HTMLInputElement>;
+  style?: React.CSSProperties;
 } & FilteredHTMLAttributes<FieldsetHTMLAttributes<HTMLFieldSetElement>, 'onChange'> &
   CheckboxGroupBaseProps<T> &
   SharedProps;
@@ -33,11 +35,13 @@ export type CheckboxGroupProps<T extends string> = {
 const CheckboxGroupWithRef = forwardRef(function CheckboxGroupWithRef<T extends string>(
   {
     children,
+    className,
     label,
     'aria-labelledby': ariaLabelledby,
     selectedValues,
     onChange,
     name,
+    style,
     testID,
     ...restProps
   }: CheckboxGroupProps<T>,
@@ -72,7 +76,13 @@ const CheckboxGroupWithRef = forwardRef(function CheckboxGroupWithRef<T extends 
   // TODO (hannah): Update default styles once Caroline has the design ready. (Add default distance between
   // checkboxes.)
   return (
-    <fieldset ref={ref} className={checkboxStyles} data-testid={testID} {...restProps}>
+    <fieldset
+      ref={ref}
+      className={cx(checkboxStyles, className)}
+      data-testid={testID}
+      {...restProps}
+      style={style}
+    >
       {label}
       {optionCheckboxes}
     </fieldset>