+++ packages/web/CHANGELOG.md
@@ -8,7 +8,11 @@ All notable changes to this project will be documented in this file!
 
 <!-- template-start -->
 
-## Unreleased
+## 7.7.0 (2/11/2025 PST)
+
+#### 🚀 Updates
+
+- Add styling props to CheckboxGroup.
 
 #### 📘 Misc